+++ seqlog/structured_logging.py
@@ -17,7 +17,6 @@ from queue import Queue
 import requests
 
 from seqlog.consumer import QueueConsumer
-from seqlog.feature_flags import FeatureFlag, is_feature_enabled
 
 # Well-known keyword arguments used by the logging system.
 _well_known_logger_kwargs = {"extra", "exc_info", "func", "sinfo"}
@@ -189,23 +188,25 @@ class StructuredLogRecord(logging.LogRecord):
             return self.msg
 
 
+def _override_root_logger():
+    """
+    Override the root logger with a `StructuredRootLogger`.
+    """
+    logging.root = StructuredRootLogger(logging.WARNING)
+    logging.Logger.root = logging.root
+    logging.Logger.manager = logging.Manager(logging.Logger.root)
+
+
+class BaseStructuredLogHandler(logging.Handler):
+    def __init__(self, level=logging.NOTSET, *args, support_extra_properties=False, **kwargs):
+        logging.Handler.__init__(self, level=level)
+        self.support_extra_properties = support_extra_properties
+
+
 class StructuredLogger(logging.Logger):
     """
     Custom (dummy) logger that understands named log arguments.
     """
-
-    def __init__(self, name, level=logging.NOTSET):
-        """
-        Create a new StructuredLogger
-        :param name: The logger name.
-        :param level: The logger minimum level (severity).
-        """
-
-        super().__init__(name, level)
-
-    @property
-    def _support_extra_properties(self):
-        return is_feature_enabled(FeatureFlag.EXTRA_PROPERTIES)
 
     def _log(self, level, msg, args, exc_info=None, extra=None, stack_info=False, **kwargs):
         """
@@ -214,7 +215,7 @@ class StructuredLogger(logging.Logger):
         :param msg: The log message or message template.
         :param args: Ordinal arguments for the message format template.
         :param exc_info: Exception information to be included in the log entry.
-        :param extra: Extra information to be included in the log entry.
+        :param extra: Extra information to be included in the log entry. Note that this will take precedence over kwargs.
         :param stack_info: Include stack-trace information in the log entry?
         :param kwargs: Keyword arguments (if any) passed to the public logger method that called _log.
         """
@@ -235,7 +236,7 @@ class StructuredLogger(logging.Logger):
 
             log_props[prop] = kwargs[prop]
 
-        if extra and self._support_extra_properties:
+        if extra:
             for extra_prop in extra.keys():
                 log_props['Extra_' + extra_prop] = extra[extra_prop]
 
@@ -334,16 +335,19 @@ class StructuredRootLogger(logging.RootLogger):
         return super().makeRecord(name, level, fn, lno, msg, args, exc_info, func, extra, sinfo)
 
 
-class ConsoleStructuredLogHandler(logging.Handler):
-    def __init__(self):
-        super().__init__()
+class ConsoleStructuredLogHandler(BaseStructuredLogHandler):
+
+    def __init__(self, *args, use_stdout=True, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.use_stdout = use_stdout
 
     def emit(self, record):
         msg = self.format(record)
 
-        print(msg)
+        out = sys.stdout if self.use_stdout else sys.stderr
+        out.write(msg)
         if hasattr(record, 'kwargs'):
-            print("\tLog entry properties: {}".format(repr(record.kwargs)))
+            out.write("\tLog entry properties: {}\n".format(repr(record.kwargs)))
 
 
 def best_effort_json_encode(arg):
@@ -367,12 +371,14 @@ def best_effort_json_encode(arg):
         return arg
 
 
-class SeqLogHandler(logging.Handler):
+class SeqLogHandler(BaseStructuredLogHandler):
     """
     Log handler that posts to Seq.
     """
 
-    def __init__(self, server_url, api_key=None, batch_size=10, auto_flush_timeout=None, json_encoder_class=None):
+    def __init__(self, server_url, api_key=None, batch_size=10, auto_flush_timeout=None, json_encoder_class=None,
+                 use_clef: bool = False, ignore_seq_submission_errors: bool = False, support_stack_info: bool = False,
+                 support_extra_properties: bool = False, **kwargs):
         """
         Create a new `SeqLogHandler`.
 
@@ -382,14 +388,26 @@ class SeqLogHandler(logging.Handler):
         :param auto_flush_timeout: If specified, the time (in seconds) before
                                    the current batch is automatically flushed.
         :param json_encoder_class: The custom JSON encoder class (or fully-qualified class name), if any, to use.
+        :param use_clef: whether to use the CLEF format
+        :param ignore_seq_submission_errors: whether to ignore failures to send logs to Seq
+        :param support_stack_info: whether to recognize stack_info as an alternative to exc_info (with some limitations)
+        :param support_extra_properties: whether to natively recognize kwargs (if True) or supply them as extra
+                                        (if False).
         """
 
-        super().__init__()
+        super().__init__(support_extra_properties=support_extra_properties, **kwargs)
 
-        self.base_server_url = server_url
-        if not self.base_server_url.endswith("/"):
-            self.base_server_url += "/"
+        server_url = server_url
+        if not server_url.endswith("/"):
+            server_url += "/"
 
+        if use_clef:
+            self.server_url = server_url + 'ingest/clef'
+        else:
+            self.server_url = server_url + 'api/events/raw'
+
+        self.support_stack_info = support_stack_info
+        self.ignore_seq_submission_errors = ignore_seq_submission_errors
         self.session = requests.Session()
 
         if api_key:
@@ -397,7 +415,7 @@ class SeqLogHandler(logging.Handler):
 
         json_encoder_class = json_encoder_class or json.encoder.JSONEncoder
         self.json_encoder_class = _ensure_class(json_encoder_class, compatible_class=json.encoder.JSONEncoder)
-
+        self.use_clef = use_clef
         self.log_queue = Queue()
         self.consumer = QueueConsumer(
             name="SeqLogHandler",
@@ -407,24 +425,6 @@ class SeqLogHandler(logging.Handler):
             auto_flush_timeout=auto_flush_timeout
         )
         self.consumer.start()
-
-    @property
-    def server_url(self):
-        if self._use_clef:
-            return self.base_server_url + 'ingest/clef'
-        return self.base_server_url + 'api/events/raw'
-
-    @property
-    def _use_clef(self):
-        return is_feature_enabled(FeatureFlag.USE_CLEF)
-
-    @property
-    def _support_stack_info(self):
-        return is_feature_enabled(FeatureFlag.STACK_INFO)
-
-    @property
-    def _ignore_seq_submission_errors(self):
-        return is_feature_enabled(FeatureFlag.IGNORE_SEQ_SUBMISSION_ERRORS)
 
     def flush(self):
         try:
@@ -480,7 +480,7 @@ class SeqLogHandler(logging.Handler):
                 continue
             processed_records.append(resp)
 
-        if self._use_clef:
+        if self.use_clef:
             request_body_json = '\r\n'.join(processed_records)
         else:
             request_body_json = '{"Events": [%s]}' % (','.join(processed_records), )
@@ -491,12 +491,12 @@ class SeqLogHandler(logging.Handler):
             response = self.session.post(
                 self.server_url,
                 data=request_body_json,
-                headers={'Content-Type': "application/vnd.serilog.clef" if self._use_clef else 'application/json'},
+                headers={'Content-Type': "application/vnd.serilog.clef" if self.use_clef else 'application/json'},
                 stream=True  # prevent '362'
             )
             response.raise_for_status()
         except requests.RequestException as requestFailed:
-            if not self._ignore_seq_submission_errors:
+            if not self.ignore_seq_submission_errors:
                 # Only notify for the first record in the batch, or we'll be generating too much noise.
                 self.handleError(batch[0])
 
@@ -528,7 +528,7 @@ class SeqLogHandler(logging.Handler):
             _callback_on_failure(exception)
 
     def _build_event_data(self, record):
-        if self._use_clef:
+        if self.use_clef:
             return self._build_event_data_clef(record)
         else:
             return self._build_event_data_ingest(record)
@@ -577,12 +577,12 @@ class SeqLogHandler(logging.Handler):
         if record.exc_text:
             # Rendered exception has already been cached
             event_data["Exception"] = record.exc_text
-        elif self._support_stack_info and record.stack_info and not record.exc_info:
+        elif self.support_stack_info and record.stack_info and not record.exc_info:
             # Feature flag is set: fall back to stack_info (sinfo) if exc_info is not present
             event_data["Exception"] = record.stack_info
         elif isinstance(record.exc_info, tuple):
             # Exception info is present
-            if record.exc_info[0] is None and self._support_stack_info and record.stack_info:
+            if record.exc_info[0] is None and self.support_stack_info and record.stack_info:
                 event_data["Exception"] = "{0}--NoException\n{1}".format(logging.getLevelName(record.levelno), record.stack_info)
             else:
                 event_data["Exception"] = record.exc_text = self.formatter.formatException(record.exc_info)
@@ -645,12 +645,12 @@ class SeqLogHandler(logging.Handler):
         if record.exc_text:
             # Rendered exception has already been cached
             event_data["@x"] = record.exc_text
-        elif self._support_stack_info and record.stack_info and not record.exc_info:
+        elif self.support_stack_info and record.stack_info and not record.exc_info:
             # Feature flag is set: fall back to stack_info (sinfo) if exc_info is not present
             event_data["@x"] = record.stack_info
         elif isinstance(record.exc_info, tuple):
             # Exception info is present
-            if record.exc_info[0] is None and self._support_stack_info and record.stack_info:
+            if record.exc_info[0] is None and self.support_stack_info and record.stack_info:
                 event_data["@x"] = "{0}--NoException\n{1}".format(logging.getLevelName(record.levelno), record.stack_info)
             else:
                 event_data["@x"] = record.exc_text = self.formatter.formatException(record.exc_info)