+++ HISTORY.rst
@@ -2,6 +2,15 @@
 History
 =======
 
+0.5.0 (2024-12-20)
+--------------------
+
+* made the configuration a little bit more Pythonic
+* FeatureFlags were removed
+* `configure_from_dict` is the official way to configure Seq
+* `support_stack_info` was removed from `log_to_console`
+* you can choose whether ConsoleStructuredLogHandler will log to stdout or to stderr.
+
 0.4.2 (2025-07-26)
 ------------------
 
@@ -14,6 +23,7 @@ History
 
 0.4.0 (2024-12-08)
 ------------------
+
 
 * You can enable and disable all of the feature flags at runtime
 * Added support for the `CLEF submission format <https://docs.datalust.co/docs/posting-raw-events>`_.